Feat: Database error fixed

+++ server/migrations/20230729181138_create_alunos_table.js
@@ -6,7 +6,7 @@ exports.up = function (knex) {
   return knex.schema.createTable('aluno', (table) => {
     table.bigIncrements('id').primary();
     table.string('nome');
-    table.integer('cpf');
+    table.bigInteger('cpf');
   });
 };
